requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.5
    # via langchain
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==42.0.5
    # via -r requirements.in
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via fastapi
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.16
    # via fastapi
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.27.0
    # via
    #   -r requirements.in
    #   fastapi
    #   langsmith
    #   openai
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via fastapi
jiter==0.12.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.5
    # via -r requirements.in
langchain-core==0.2.7
    # via
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.7
    # via -r requirements.in
langchain-text-splitters==0.2.0
    # via langchain
langsmith==0.1.147
    # via
    #   langchain
    #   langchain-core
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via langchain
openai==1.109.1
    # via langchain-openai
orjson==3.11.4
    # via
    #   fastapi
    #   langsmith
packaging==24.2
    # via langchain-core
propcache==0.4.1
    # via yarl
pycparser==2.23
    # via cffi
pydantic==2.7.1
    # via
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.18.2
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
pygments==2.19.2
    # via rich
pyjwt==2.8.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
    #   uvicorn
redis==5.0.4
    # via -r requirements.in
regex==2025.11.3
    # via tiktoken
requests==2.32.5
    # via
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.17.0
    # via fastapi-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   httpx
    #   openai
sqlalchemy==2.0.29
    # via langchain
starlette==0.37.2
    # via fastapi
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-core
tiktoken==0.7.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typer==0.20.0
    # via fastapi-cli
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
ujson==5.11.0
    # via fastapi
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.29.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
yarl==1.22.0
    # via aiohttp
